Rename USPS classes to consistent casing and update references
