--- careconnect2025/backend/core/src/main/java/com/careconnect/model/UspsDigest.java
@@ -1,10 +1,0 @@
-package com.careconnect.model;
-
-import java.time.OffsetDateTime;
-import java.util.List;
-
-public record USPSDigest(
-        OffsetDateTime digestDate,
-        List<MailPiece> mailpieces,
-        List<PackageItem> packages
-) {}
